moved user init test to sys prgrms
