+++ light-system-programs/tests/user.tests.ts
@@ -1,9 +1,11 @@
 import { Keypair, Keypair as SolanaKeypair } from "@solana/web3.js";
 import _ from "lodash";
+import { sign } from "tweetnacl";
 const chai = require("chai");
 const chaiAsPromised = require("chai-as-promised");
 // init chai-as-promised support
 chai.use(chaiAsPromised);
+
 import { bs58 } from "@coral-xyz/anchor/dist/cjs/utils/bytes";
 let circomlibjs = require("circomlibjs");
 import {
@@ -25,10 +27,11 @@ import {
   ConfirmOptions,
   airdropShieldedSol,
   TOKEN_ACCOUNT_FEE,
+  useWallet,
 } from "@lightprotocol/zk.js";
 
 import { BN, AnchorProvider, setProvider } from "@coral-xyz/anchor";
-import { assert } from "chai";
+import { assert, expect } from "chai";
 
 var POSEIDON;
 var RELAYER: TestRelayer, provider: Provider, user: User;
@@ -80,6 +83,63 @@ describe("Test User", () => {
     });
 
     user = await User.init({ provider });
+  });
+
+  it("externally supplied seed vs internal seed (user derivation)", async () => {
+    const message =
+      "IMPORTANT:\nThe application will be able to spend \nyour shielded assets. \n\nOnly sign the message if you trust this\n application.\n\n View all verified integrations here: \n'https://docs.lightprotocol.com/partners'";
+
+    const walletMock = useWallet(ADMIN_AUTH_KEYPAIR);
+    const walletMock2 = useWallet(Keypair.generate());
+
+    const encodedMessage = new TextEncoder().encode(message);
+    const signature = await walletMock.signMessage(encodedMessage);
+    const signature2 = await walletMock2.signMessage(encodedMessage);
+
+    if (
+      !sign.detached.verify(
+        encodedMessage,
+        signature,
+        walletMock.publicKey.toBytes(),
+      )
+    )
+      throw new Error("Invalid signature!");
+
+    let testRelayer = new TestRelayer({
+      relayerPubkey: ADMIN_AUTH_KEYPAIR.publicKey,
+      relayerRecipientSol: Keypair.generate().publicKey,
+      relayerFee: new BN(100_000),
+      payer: ADMIN_AUTH_KEYPAIR,
+    });
+    const providerExternalSeed = await Provider.init({
+      relayer: testRelayer,
+      wallet: walletMock,
+      confirmConfig,
+    });
+    const providerInternalSeed = await Provider.init({
+      relayer: testRelayer,
+      wallet: walletMock,
+      confirmConfig,
+    });
+
+    const userExternal = await User.init({
+      provider: providerExternalSeed,
+      seed: bs58.encode(signature),
+    });
+    const userExternal2 = await User.init({
+      provider: providerExternalSeed,
+      seed: bs58.encode(signature2),
+    });
+    const userInternal = await User.init({
+      provider: providerInternalSeed,
+    });
+
+    const externalKey = await userExternal.account.getPublicKey();
+    const externalKey2 = await userExternal2.account.getPublicKey();
+    const internalKey = await userInternal.account.getPublicKey();
+
+    expect(externalKey).to.deep.equal(internalKey);
+    expect(externalKey2).to.not.deep.equal(internalKey);
   });
 
   it("(user class) shield SPL", async () => {
