+++ light-zk.js/package.json
@@ -17,7 +17,6 @@
     "test-provider": "ts-mocha --resolveJsonModule ./tsconfig.json -t 100000000 tests/provider.test.ts --exit",
     "test-selectInUtxos": "ts-mocha --resolveJsonModule ./tsconfig.json -t 100000000 tests/selectInUtxos.test.ts --exit",
     "test-balance": "ts-mocha --resolveJsonModule ./tsconfig.json -t 100000000 tests/balance.test.ts --exit",
-    "test-user": "ts-mocha --resolveJsonModule ./tsconfig.json -t 100000000 tests/user.test.ts --exit",
     "test-convertDecimals": "ts-mocha --resolveJsonModule ./tsconfig.json -t 100000000 tests/convertDecimals.test.ts --exit",
     "build": "yarn tsc",
     "format": "prettier --write \"src/**/*.{ts,js}\" \"tests/**/*.{ts,js}\" -w",
